+++ monkey/monkey_island/cc/server_utils/encryption/__init__.py
@@ -1,18 +1,17 @@
-from monkey_island.cc.server_utils.encryption.i_encryptor import IEncryptor
-from monkey_island.cc.server_utils.encryption.key_based_encryptor import KeyBasedEncryptor
-from monkey_island.cc.server_utils.encryption.password_based_string_encryptior import (
+from monkey_island.cc.server_utils.encryption.encryptors.i_encryptor import IEncryptor
+from monkey_island.cc.server_utils.encryption.encryptors.key_based_encryptor import (
+    KeyBasedEncryptor,
+)
+from monkey_island.cc.server_utils.encryption.encryptors.password_based_string_encryptior import (
     PasswordBasedStringEncryptor,
     is_encrypted,
 )
-from .encryptor_factory import (
-    FactoryNotInitializedError,
-    remove_old_datastore_key,
-    get_encryptor_factory,
-    _get_secret_from_credentials,
-    initialize_encryptor_factory,
+from monkey_island.cc.server_utils.encryption.encryptors.password_based_bytes_encryption import (
+    PasswordBasedBytesEncryptor,
+    InvalidCredentialsError,
+    InvalidCiphertextError,
 )
 from .data_store_encryptor import initialize_datastore_encryptor, get_datastore_encryptor
-from .password_based_bytes_encryption import InvalidCredentialsError, InvalidCiphertextError
 from .dict_encryption.dict_encryptor import (
     SensitiveField,
     encrypt_dict,
