+++ monkey/monkey_island/cc/server_utils/encryption/encryptors/i_encryptor.py
