+++ monkey/monkey_island/cc/server_utils/encryption/encryptors/key_based_encryptor.py
