+++ monkey/monkey_island/cc/server_utils/encryption/encryptors/password_based_bytes_encryption.py
